js callback and sorting arrays
the goal of the exercise is to learn what I thought so far more complicated functions and arrays. They turned out to be quite easy, but still require a lot of work to master them in their entirety.

Exercise does not reflect the size of my work at all because it is constantly deleting and adding new lines of code. Pushes them naghithub to easily find them in the future and to prolong the strike